+++ frontend/src/pages/game/Models/GameCanvas.jsx
@@ -11,7 +11,7 @@ import { Canvas } from '@react-three/fiber';
 import { EffectComposer, Bloom } from '@react-three/postprocessing';
 import { Html, useProgress } from '@react-three/drei';
 import { setCanvasLoaded } from '../../../store/reducers/gameSlice';
-import ShipModel from './ShipModel';
+import ShipModel from './Ship/ShipModel/ShipModel';
 import BackgroundModel from './BackgoundModel';
 import MoveCamera from './MoveCamera';
 import FogModel from './FogModel';
